requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.8
openai>=1.6.1,<2.0.0
supabase>=2.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.24.0,<0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0 